+++ src/simulator/units.rs
@@ -1,0 +1,4 @@
+pub fn celsius2kelvin(deg_celsius: f32) -> f32 {
+    // Convert degrees C to Kelvin
+    deg_celsius + 273.15
+}
